+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `laravel-trend` will be documented in this file.
 
+## 1.0.1 - 2025-04-18
+
+- Fixed SQLite week interval date format compatibility issue
+- Added support for SQLite 3.46+ ISO week format
+
 ## 1.0.0 - 202X-XX-XX
 
 - initial release